Less memory allocations & more null move conditions

+++ Renegade/Engine.cpp
@@ -229,7 +229,7 @@ void Engine::Start() {
 			}
 
 			SearchParams params;
-			for (int i = 1; i < parts.size(); i++) {
+			for (int64_t i = 1; i < parts.size(); i++) {
 				if (parts[i] == "wtime") { params.wtime = stoi(parts[i + 1]); i++; }
 				if (parts[i] == "btime") { params.btime = stoi(parts[i + 1]); i++; }
 				if (parts[i] == "movestogo") { params.movestogo = stoi(parts[i + 1]); i++; }
